mcp/requirements.txt
mcp>=1.0.0
fastembed>=0.4.0
sqlite-vec>=0.1.6
python-frontmatter>=1.1.0
numpy>=1.24.0
